Revert "super weird toggle bug"
This reverts commit ec26b1eb737d21bc5554af384c906e2708a92af4.
