+++ ngapp/src/app/student/dashboard/dashboard.component.html
@@ -89,10 +89,10 @@
   >
     <mat-slide-toggle
       class="example-margin"
-      *ngIf="synthButtons"
       [color]="primary"
-      [checked]="synthButtons.checked()"
-      (click)="synthButtons.toggle()"
+      [checked]="checked"
+      [disabled]="disabled"
+      (click)="toggleSynthesisPlayButtons()"
     >
     </mat-slide-toggle>
     <span (click)="toggleRightDrawer('synthesis')">
